requirements.txt
openai==0.28
deta
streamlit_option_menu==0.3.6
json
datatime